Transplant codes to Memo component
